requirements.txt
pandas
python-dotenv
requests
mysql-connector-python
pydantic
pyspark
